fix(ci): python-full-coverage — split residue tests under 1500-line limit
Shard 8 failed test_first_party_code_files_stay_under_line_limit because
the post-operation residue module grew to 1982 lines after coverage tests.
