+++ tests/unit/runtime/test_pr_monitor_pre_push_validation_post_operation_residue_parts/conftest.py
@@ -1,0 +1,18 @@
+"""Fixtures for post-operation residue cleanup regression tests."""
+
+from __future__ import annotations
+
+from collections.abc import AsyncIterator
+
+import pytest
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
+
+from awf.db.session import make_session_factory
+from tests.postgres import postgres_test_engine
+
+
+@pytest.fixture
+async def factory() -> AsyncIterator[async_sessionmaker[AsyncSession]]:
+    """Yield a scoped async SQLAlchemy session factory for tests."""
+    async with postgres_test_engine() as engine:
+        yield make_session_factory(engine)
